src/ingestion/consumer/requirements.txt
boto3==1.40.55
kafka-python==2.0.2
python-decouple==3.8
pandas==3.0.2
pyarrow==21.0.0
scikit-learn==1.7.1
